miles

+++ WorkiSiteWeb/Worki.Data/Models/SearchCriteria.cs
@@ -45,8 +45,8 @@ namespace Worki.Data.Models
             toRet[MiscHelpers.SeoConstants.Page] = page;
             toRet[MiscHelpers.SeoConstants.Place] = Place;
             toRet[MiscHelpers.SeoConstants.OfferType] = Localisation.GetSeoStringOfferFromType(OfferData.Type);
-            toRet[MiscHelpers.SeoConstants.Latitude] = LocalisationData.Latitude;
-            toRet[MiscHelpers.SeoConstants.Longitude] = LocalisationData.Longitude;
+            toRet[MiscHelpers.SeoConstants.Latitude] = (float)LocalisationData.Latitude;
+            toRet[MiscHelpers.SeoConstants.Longitude] = (float)LocalisationData.Longitude;
             toRet[MiscHelpers.SeoConstants.PlaceName] = LocalisationData.Name;
 
             toRet[MiscHelpers.SeoConstants.Order] = (int)OrderBy;
